Move calendar as seperate component
